requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cloudinary==1.30.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
psycopg2==2.9.4
PyJWT==1.7.1
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
